feat : 쉐어 작성자는 참가 취소를 할 수 없도록 구현

+++ src/main/java/louie/hanse/shareplate/service/EntryService.java
@@ -51,8 +51,8 @@ public class EntryService {
 
     @Transactional
     public void cancel(Long shareId, Long memberId) {
-        memberService.findByIdOrElseThrow(memberId);
-        Share share = shareService.findByIdOrElseThrow(shareId);
+        Member member = memberService.findByIdOrElseThrow(memberId);
+        Share share = shareService.findWithWriterByIdOrElseThrow(shareId);
 
         share.isCanceledThrowException();
         if (!isExistEntry(shareId, memberId)) {
@@ -63,6 +63,9 @@ public class EntryService {
         }
         if (share.isLeftLessThanAnHour()) {
             throw new GlobalException(EntryExceptionType.CLOSE_TO_THE_CLOSED_DATE_TIME);
+        }
+        if (share.isWriter(member)) {
+            throw new GlobalException(EntryExceptionType.SHARE_WRITER_CANNOT_ENTRY_CANCEL);
         }
         entryRepository.deleteByMemberIdAndShareId(memberId, shareId);
     }
